--- sources/src/main/java/com/ullink/slack/simpleslackapi/impl/GenericSlackReplyImpl.java
@@ -1,20 +1,0 @@
-package com.ullink.slack.simpleslackapi.impl;
-
-import com.ullink.slack.simpleslackapi.replies.GenericSlackReply;
-
-class GenericSlackReplyImpl implements GenericSlackReply
-{
-    private String obj;
-
-    public GenericSlackReplyImpl(String obj)
-    {
-        this.obj = obj;
-    }
-
-    @Override
-    public String getPlainAnswer()
-    {
-        return obj;
-    }
-
-}
