+++ sources/src/main/java/com/ullink/slack/simpleslackapi/replies/EmojiSlackReply.java
@@ -2,8 +2,22 @@ package com.ullink.slack.simpleslackapi.replies;
 
 import java.util.Map;
 
-public interface EmojiSlackReply extends SlackReply {
-    String getTimestamp();
+public class EmojiSlackReply extends SlackReplyImpl implements SlackReply {
 
-    Map<String,String> getEmojis();
+    private String timestamp;
+    private Map<String, String>  emoji;
+
+    public EmojiSlackReply(boolean ok, String error, Map<String, String>  emoji, String timestamp) {
+        super(ok, error);
+        this.timestamp = timestamp;
+        this.emoji = emoji;
+    }
+
+    public String getTimestamp() {
+        return timestamp;
+    }
+
+    public Map<String, String> getEmojis() {
+        return emoji;
+    }
 }
