+++ sources/src/main/java/com/ullink/slack/simpleslackapi/replies/SlackChannelReply.java
@@ -2,6 +2,17 @@ package com.ullink.slack.simpleslackapi.replies;
 
 import com.ullink.slack.simpleslackapi.SlackChannel;
 
-public interface SlackChannelReply extends ParsedSlackReply {
-    SlackChannel getSlackChannel();
+public class SlackChannelReply extends SlackReplyImpl implements ParsedSlackReply {
+    private SlackChannel slackChannel;
+
+    SlackChannelReply(boolean ok, String error, SlackChannel slackChannel)
+    {
+        super(ok,error);
+        this.slackChannel = slackChannel;
+    }
+
+    public SlackChannel getSlackChannel()
+    {
+        return slackChannel;
+    }
 }
