+++ sources/src/main/java/com/ullink/slack/simpleslackapi/replies/SlackMessageReply.java
@@ -1,7 +1,24 @@
 package com.ullink.slack.simpleslackapi.replies;
 
-public interface SlackMessageReply extends ParsedSlackReply {
-    long getReplyTo();
-    String getTimestamp();
+//TODO: figure out this hierarchy
+public class SlackMessageReply extends SlackReplyImpl implements ParsedSlackReply {
+    private long replyTo;
+    private String timestamp;
 
+    SlackMessageReply(boolean ok, String error, long replyTo, String timestamp)
+    {
+        super(ok, error);
+        this.replyTo = replyTo;
+        this.timestamp = timestamp;
+    }
+
+    public long getReplyTo()
+    {
+        return replyTo;
+    }
+
+    public String getTimestamp()
+    {
+        return timestamp;
+    }
 }
